refactor: 불필요한 메서드 및 예외 삭제
