+++ src/main/java/roomescape/controller/AdminLoginInterceptor.java
@@ -1,23 +1,30 @@
-package roomescape.config;
+package roomescape.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
+import roomescape.config.JwtTokenProvider;
+import roomescape.domain.Role;
 
 @Component
-public class MemberLoginInterceptor implements HandlerInterceptor {
+public class AdminLoginInterceptor implements HandlerInterceptor {
 
     private final JwtTokenProvider jwtTokenProvider;
 
-    public MemberLoginInterceptor(final JwtTokenProvider jwtTokenProvider) {
+    public AdminLoginInterceptor(final JwtTokenProvider jwtTokenProvider) {
         this.jwtTokenProvider = jwtTokenProvider;
     }
 
     @Override
-    public boolean preHandle(final HttpServletRequest request, final HttpServletResponse response, final Object handler)
-            throws Exception {
+    public boolean preHandle(final HttpServletRequest request, final HttpServletResponse response, final Object handler) {
         if (jwtTokenProvider.doesNotRequestHasCookie(request) || jwtTokenProvider.doesNotRequestHasToken(request)) {
+            response.setStatus(401);
+            return false;
+        }
+
+        Role role = jwtTokenProvider.extractRole(request);
+        if (Role.ADMIN != role) {
             response.setStatus(401);
             return false;
         }
